add security config JWT not finish

+++ src/main/java/com/mj/market/app/user/UserService.java
@@ -1,8 +1,10 @@
 package com.mj.market.app.user;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
-import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -12,39 +14,23 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class UserService implements UserDetailsService {
+@RequiredArgsConstructor
+public class UserService {
 
     private final UserRepository userRepository;
-
-    @Autowired
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
+    private final AuthenticationManager authenticationManager;
+    private final JwtService jwtService;
+    private final PasswordEncoder passwordEncoder;
 
     public User findByUsername(String username) throws UsernameNotFoundException {
         return userRepository.findByUsername(username);
     }
-
     public List<User> getUsers() {
         return userRepository.findAll();
     }
 
-    //For Security
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return userRepository.findByUsername(username);
-    }
-
-    @Bean
-    public static PasswordEncoder passwordEncoder(){
-        return new BCryptPasswordEncoder(10);
-    }
-
     public void saveUser(User user) {
-        user.setPassword(UserService.passwordEncoder().encode(user.getPassword()));
-        user.setRole("ROLE_MODERATOR");
-        userRepository.save(user);
+        if(findByUsername(user.getUsername()) == null) userRepository.save(user);
     }
     public User findById(Long id) throws UsernameNotFoundException{
         return userRepository.findById(id).get();
@@ -53,11 +39,7 @@ public class UserService implements UserDetailsService {
     public void deleteUser(Long id) throws UsernameNotFoundException {
         //initialization Admin ID cant be deleted
         Long initAdminId = Long.valueOf(1);
-        if(id != initAdminId) {
-            {
-                userRepository.deleteById(id);
-            }
-        }
+        if(id != initAdminId) userRepository.deleteById(id);
     }
 
     public void updateUser(User user) {
@@ -68,6 +50,32 @@ public class UserService implements UserDetailsService {
     }
 
     public User findByEmail(String email) {
-        return userRepository.findByEmail(email);
+        return userRepository.findByEmail(email).orElseThrow(
+                ()-> new UsernameNotFoundException("No user with this email"));
+    }
+
+    public AuthenticationResponse authenticate(AuthenticationRequest request) {
+        authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(
+                        request.getEmail(),
+                        request.getPassword()
+                )
+        );
+        var user = userRepository.findByEmail(request.getEmail()).orElseThrow(
+                ()-> new UsernameNotFoundException("No user with this email"));
+        var token = jwtService.generateToken(user);
+        return AuthenticationResponse.builder().token(token).build();
+    }
+
+    public AuthenticationResponse register(RegisterRequest request) {
+        var user = User.builder()
+                .username(request.getName())
+                .email(request.getName())
+                .password(passwordEncoder.encode(request.getPassword()))
+                .role(Role.USER)
+                .build();
+        userRepository.save(user);
+        var token = jwtService.generateToken(user);
+        return AuthenticationResponse.builder().token(token).build();
     }
 }
